Split storage provider into multiple files
