+++ backend/requirements.txt
@@ -25,6 +25,7 @@ bcrypt==4.2.0
 pymongo
 redis
 boto3==1.35.0
+aioboto3==13.2.0
 
 argon2-cffi==23.1.0
 APScheduler==3.10.4
